some fixes after mypy
