+++ pyproject.toml
@@ -22,6 +22,8 @@ isort = "^5.12.0"
 pytest-mock = "^3.14.0"
 pytest-asyncio = "^0.23.7"
 pytest-httpx = "^0.30.0"
+pre-commit = "^3.8.0"
+types-beautifulsoup4 = "^4.12.0.20240511"
 
 [build-system]
 requires = ["poetry-core"]
