+++ tests/integration/test_site.py
@@ -2,7 +2,7 @@ from urllib.parse import urljoin
 
 import pytest
 
-from dataservice.client import HttpXClient
+from httpx_client import HttpXClient
 from dataservice.models import Request, Response
 from dataservice.service import DataService
 import dataservice.logging_config
